Gc7.0

+++ server/java-server/src/main/java/org/fatmansoft/teach/models/Course.java
@@ -43,7 +43,7 @@ public class Course implements Serializable {
     private String name;
     @Setter
     @Getter
-    private Integer credit;
+    private Double credit;
 
     @Getter
     @Setter
